Phpstan Baseline Fixes 2022-09-21 (#3080)
* Phpstan Baseline Fixes 2022-09-21

Eliminate about 200 more lines from Phpstan baseline. For Helper/Sample and Helper/Html and others, many properties are declared as protected despite the fact that the classes do not extend any other class, and there are no classes which extend them. They are changed to private; this could be a breaking change in circumstances for which I cannot think of a use case (user extends class for some reason).

* Slightly Botched Merge Commit

Hope this fixes phpstan.

+++ src/PhpSpreadsheet/Helper/Sample.php
@@ -77,6 +77,11 @@ class Sample
     {
         // Populate samples
         $baseDir = realpath(__DIR__ . '/../../../samples');
+        if ($baseDir === false) {
+            // @codeCoverageIgnoreStart
+            throw new RuntimeException('realpath returned false');
+            // @codeCoverageIgnoreEnd
+        }
         $directory = new RecursiveDirectoryIterator($baseDir);
         $iterator = new RecursiveIteratorIterator($directory);
         $regex = new RegexIterator($iterator, '/^.+\.php$/', RecursiveRegexIterator::GET_MATCH);
@@ -84,6 +89,11 @@ class Sample
         $files = [];
         foreach ($regex as $file) {
             $file = str_replace(str_replace('\\', '/', $baseDir) . '/', '', str_replace('\\', '/', $file[0]));
+            if (is_array($file)) {
+                // @codeCoverageIgnoreStart
+                throw new RuntimeException('str_replace returned array');
+                // @codeCoverageIgnoreEnd
+            }
             $info = pathinfo($file);
             $category = str_replace('_', ' ', $info['dirname'] ?? '');
             $name = str_replace('_', ' ', (string) preg_replace('/(|\.php)/', '', $info['filename']));
@@ -172,12 +182,17 @@ class Sample
     public function getTemporaryFilename($extension = 'xlsx')
     {
         $temporaryFilename = tempnam($this->getTemporaryFolder(), 'phpspreadsheet-');
+        if ($temporaryFilename === false) {
+            // @codeCoverageIgnoreStart
+            throw new RuntimeException('tempnam returned false');
+            // @codeCoverageIgnoreEnd
+        }
         unlink($temporaryFilename);
 
         return $temporaryFilename . '.' . $extension;
     }
 
-    public function log($message): void
+    public function log(string $message): void
     {
         $eol = $this->isCli() ? PHP_EOL : '<br />';
         echo date('H:i:s ') . $message . $eol;
